finished threaded CSP functions - still might change depending on how I want workload to be distributed

+++ ThreadedSolver/CSP/assign.c
@@ -1,22 +1,22 @@
 #include "CSP.h"
 
-void assign(ushort variable, ushort value)
+void assign(uint8_t variable, uint8_t value)
 {
   assignments[variable] = value;
   csp.nassigns++;
 }
 
-void unassign(ushort variable)
+void unassign(uint8_t variable)
 {
   assignments[variable] = 0;
 }
 
-ushort nconflicts(ushort variable, ushort value)
+uint8_t nconflicts(uint8_t variable, uint8_t value)
 {
-  ushort neighbor, n_idx, conflicts = 0;
+  uint8_t neighbor, n_idx, conflicts = 0;
   for (n_idx = 0; n_idx < NUM_NEIGHBORS; n_idx++) {
     neighbor = neighbors[variable][n_idx];
-    if (constraint(variable, value, neighbor, assignments[n_idx]) ==0)
+    if (constraint(variable, value, neighbor, assignments[n_idx]) == 0)
       conflicts++;
   }
   return conflicts;
